requirements.txt
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.20
gradio==4.0.0
pandas==2.0.0
numpy==1.24.0
scikit-learn==1.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
plotly==5.17.0

